Rework popular pages widget.
